requirements.txt
selenium>=3.141.0
robotframework>=3.1.2
robotframework-faker>=4.3.0
robotframework-seleniumlibrary>=4.1.0
robotframework-selenium2library>=3.0.0
robotframework-pageobjectlibrary>=1.0.2
behave>=1.2.6
requests>=2.23.0

